Added comments to index.html and added developer logs
Added some comments to index.html to places which will possibly be useful during later stages of development and to improve readability. Also added two developer logs, 'Log_WeeklyUpdates.txt', which will be updated with weekly project updates and 'Log_PageInfo.txt' which contains goals and the structure of the web page.