Added log in and sign up to index.html with popup functionality
Added two new elements to index.html, signUp and logIn. Then I wrote a short javascript code to animate them to pup up when you click on either sign up or log in. Then they are animated to go back out of view when you click again on the sign up button, log in button or when you click the little x icon in the corner of the window